requirements.txt
absl-py==0.9.0
astor==0.8.0
beautifulsoup4==4.9.3
brotlipy==0.7.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==3.1.1
future==0.18.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.16.1
h5py==2.10.0
idna==2.10
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
line-bot-sdk==1.17.0
Markdown==3.1.1
numpy==1.18.1
olefile==0.46
opt-einsum==3.1.0
Pillow==8.0.1
pycparser==2.20
pyOpenSSL==19.1.0
PySocks==1.7.1
PyYAML==5.3.1
requests==2.24.0
scipy==1.4.1
six==1.14.0
soupsieve==2.0.1
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
urllib3==1.25.11
Werkzeug==1.0.0
wrapt==1.12.1
